requirements.txt
jupyter
pandas
*.tar.gz
numpy
scikit-learn
yahoo-finance
yfinance
requests
seaborn
cvxopt
plotly
zipline-reloaded
cufflinks
chart-studio
